Improved: code for utils and added toast when notification is added
