+++ .env.example
@@ -13,3 +13,6 @@ VUE_APP_CURRENCY_FORMATS={"en": {"currency": {"style": "currency","currency": "U
 VUE_APP_RF_CNFG_MPNG={ "allowDeliveryMethodUpdate": "CUST_DLVRMTHD_UPDATE", "allowDeliveryAddressUpdate": "CUST_DLVRADR_UPDATE", "allowPickupUpdate": "CUST_PCKUP_UPDATE", "allowCancel": "CUST_ALLOW_CNCL", "shippingMethod": "RF_SHIPPING_METHOD"}
 VUE_APP_DEFAULT_LOG_LEVEL="error"
 VUE_APP_LOGIN_URL="http://launchpad.hotwax.io/login"
+VUE_APP_NOTIF_APP_ID=BOPIS
+VUE_APP_NOTIF_ENUM_TYPE_ID=NOTIF_BOPIS
+VUE_APP_FIREBASE_CONFIG={"apiKey": "AIzaSyAq-jeG6dy8y2O6p_RYKk_MpkfTZ97GF_Q","authDomain": "hotwax-digital-commerce.firebaseapp.com","databaseURL": "https://hotwax-digital-commerce.firebaseio.com","projectId": "hotwax-digital-commerce","storageBucket": "hotwax-digital-commerce.appspot.com","messagingSenderId": "211268342110","appId": "1:211268342110:web:6fa33f0d16129925c27fcf"}
