+++ src/locales/en.json
@@ -56,6 +56,7 @@
   "Logout": "Logout",
   "Mismatch": "Mismatch",
   "More": "More",
+  "New notification received.": "New notification received.",
   "No inventory details found": "No inventory details found",
   "Not in stock": "Not in stock",
   "Not in Stock": "Not in Stock",
