+++ src/locales/es.json
@@ -56,6 +56,7 @@
   "Logout": "Cerrar sesión",
   "Mismatch": "Desajuste",
   "More": "Más",
+  "New notification received.": "New notification received.",
   "No inventory details found": "No se encontraron detalles de inventario",
   "Not in stock": "No disponible en el inventario",
   "No products found": "No se encontraron productos",
