+++ src/locales/ja.json
@@ -56,6 +56,7 @@
   "Logout": "ログアウト",
   "Mismatch": "不一致",
   "More": "More",
+  "New notification received.": "New notification received.",
   "No inventory details found": "在庫の詳細が見つかりません",
   "Not in Stock": "在庫切れ",
   "No products found": "商品が見つかりません",
