+++ src/main.ts
@@ -28,11 +28,14 @@ import "@hotwax/apps-theme";
 
 import i18n from './i18n'
 import store from './store'
+
 import permissionPlugin from '@/authorization';
 import permissionRules from '@/authorization/Rules';
 import permissionActions from '@/authorization/Actions';
+
 import { dxpComponents } from '@hotwax/dxp-components'
-import { login, logout, loader } from './user-utils';
+import { login, logout, loader } from '@/utils/user';
+import { showNewNotificationToast } from '@/utils/notifications';
 import {
   getConfig,
   getNotificationEnumIds,
@@ -59,6 +62,7 @@ const app = createApp(App)
     appLoginUrl: process.env.VUE_APP_LOGIN_URL as string,
     notificationApplicationId: process.env.VUE_APP_NOTIF_APP_ID,
     notificationEnumTypeId: process.env.VUE_APP_NOTIF_ENUM_TYPE_ID,
+    appFirebaseConfig: JSON.parse(process.env.VUE_APP_FIREBASE_CONFIG),
     getConfig,
     getNotificationEnumIds,
     getNotificationUserPrefTypeIds,
@@ -67,6 +71,7 @@ const app = createApp(App)
     login,
     logout,
     removeClientRegistrationToken,
+    showNewNotificationToast,
     storeClientRegistrationToken, 
     subscribeTopic,
     unsubscribeTopic
