+++ src/utils/notifications.ts
@@ -1,0 +1,10 @@
+import { translate } from '@/i18n';
+import { showToast } from '@/utils'
+
+const showNewNotificationToast = () => {
+  showToast(translate('New notification received.'));
+}
+
+export {
+  showNewNotificationToast
+}
